add mtl parser

+++ src/main/java/com/example/acg_labs/model/Color.java
@@ -1,35 +1,35 @@
 package com.example.acg_labs.model;
 
 public class Color {
-    private float r;
-    private float g;
-    private float b;
+    private double r;
+    private double g;
+    private double b;
 
-    public float getR() {
+    public double getR() {
         return r;
     }
 
-    public void setR(float r) {
+    public void setR(double r) {
         this.r = r;
     }
 
-    public float getG() {
+    public double getG() {
         return g;
     }
 
-    public void setG(float g) {
+    public void setG(double g) {
         this.g = g;
     }
 
-    public float getB() {
+    public double getB() {
         return b;
     }
 
-    public void setB(float b) {
+    public void setB(double b) {
         this.b = b;
     }
 
-    public Color(float r, float g, float b) {
+    public Color(double r, double g, double b) {
         this.r = r;
         this.g = g;
         this.b = b;
